docker_test/requirements.txt
opencv-python==4.4.0.44
pytesseract==0.3.6
nltk==3.5
imutils==0.5.3
pyenchant==3.1.1
numpy==1.19.1
requests==2.24.0
Pillow==7.2.0
opencv-contrib-python==3.4.4.19
google-cloud-storage==1.31.0
Flask==1.1.2
gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'